qcacld: Invalidate VOS HDD context when HDD context is freed
Change "qcacld: Fix for driver crash when invalid country code is set"
attempted to fix the issue that, in cases where hdd_wlan_startup()
failed, subsequently __hdd_wlan_exit() would try to cleanup data
structures that had already been cleaned up by the failure handling
code within hdd_wlan_startup().  This change had a flaw, however, in
that it relied upon one of the data structures that was cleaned up,
which resulted in a "write after free" error.  In reality the only
change required is a simple one.  When hdd_wlan_startup() fails and
the cleanup operations are performed, invalate the HDD context pointer
held by VOSS so that when subsequent code executes it will not find
a valid context and hence will know that no cleanup is required.

Change-Id: I2f0bb943bc01f84059b26c661a7ac892d71e154e
CRs-fixed: 707864