qcacld: Enable HTT batching
Currently HTT is batching the msdus but while handingover the
msdu to HTC it is giving 1 msdu at a time. Inorder to enhance
HTC bundling HTT will indicate HTC whether the msdu is part of
batch of msdu's through more_data flag. HTT will set more_data
flag as 1 for msdu's except the last msdu in the batch of msdu's.

HTC will make use if information if HTC bundling is enabled.
HTC will queue the packet with more_data flag set. Once the
msdu_flag is set to 0 HTC will try to send the batch of msdu
by bundling the data.

This will help HTC to bundle more data and improve throughput.

Change-Id: I33d104c2a37d90282ed16b2b0c542a7f8875c6fa
CRs-Fixed: 716742