wlan: SME: add GPIO LED flashing feature
Passing parameters from user to firmware for LED flashing control.
The state of the LED GPIO control is determined by two 32 bit values
(X_0 and X_1) to produce a 64 bit value. Each 32 bit value consists
of 4 bytes, where each byte defines the number of 50ms intervals that
the GPIO will remain at a predetermined state. The 64 bit value provides
8 unique GPIO timing intervals. The pattern starts with the MSB of X_0
and continues to the LSB of X_1. After executing the timer interval of
the LSB of X_1, the pattern returns to the MSB of X_0 and repeats. The
GPIO state for each timing interval  alternates from Low to High and the
first interval of the pattern represents the time when the GPIO is Low.
When a timing interval of Zero is reached, it is skipped and moves on to
the next interval.

Change-Id: I97833f81005d61118617441aab8fd1876ab531ac
CRs-Fixed: 733909