qcacld:SDIO Fix for Data Stall issue
There is race where HTCSendDataPkt() thinks the
tx queue is in access by referring TxProcessCount so
it puts the data in tx queue and returns.

But HTCTrySend() which gets called from
HTCTxCompletionHandler() serviced the queue
but not yet cleared the TxProcessCount.
But unlocks the traget htc lock.

This race results in a case where the tx packet is not
processed at all and txrx scheduler is waiting for this
credit to download other packets.

So fix it by unlocking htc lock after clearing
TxProcessCount.

CRs-Fixed: 751626
Change-Id: If00d9e8c4757313671c58aee3d9869553424fdd6