qcacld: HDD: Connected AP rssi is constant when LCD is off
Root cause of this issue is, if the DUT LCD is off and scan is
triggered using lowi_test then during inform_bss the RSSI of the
connected AP is taken from pAdapter->rssi.

Instead of pAdapter->rssi, we should be taking the rssi from the
scan result (bss descriptor rssi).

Change-Id: Idfdb8a61bd4c2b2747172d12097693b272a88aa7
CRs-Fixed: 758339