qcacld: Fix Driver Load Failure in static build driver
With the current design driver never gets unloaded if the
driver is inbuilt into the kernel.

On SDIO based platforms, the second load is failing as the
probed state is not cleared as part of exit.

The probed state adds no value to our current design.
Hence remove it completely and fix the driver load failure
issue.

Change-Id: Ie6a0d56a306b6b1da5d8ea36d745b7c23f14b54c
CRs-Fixed: 884503